Crear Partido mejor

+++ app/src/main/java/com/example/padelhub/modelo/Usuario.kt
@@ -2,7 +2,7 @@ package com.example.padelhub.modelo
 
 data class Usuario(
     val nombre: String,
-    val edad: Int,
-    val email : String
+    val edad: Long,
+    val email: String
 )
